requirements_dev.txt
pip
bumpversion==0.5.3
wheel==0.32.1
watchdog==0.9.0
flake8==3.5.0
tox==3.5.2
coverage==4.5.1
Sphinx==1.8.1
twine==1.12.1
praw==6.1.1
PyYAML==5.1
spotipy==2.4.4


